[ch06-07] move ast node to separate file
