+++ lm/ch07/CMakeLists.txt
@@ -8,6 +8,7 @@ add_executable(lm-ch07
   ../../gtest_main.cc
   EvaluatorTest-ch07-extensions.cpp
   ShuntingYardTest.cpp
+  ASTNode.cpp
   )
 
 target_link_libraries(lm-ch07 lm-calculator-ext-lib lm-calculator-lib gtest gmock)
