+++ lm/ch07/ShuntingYardTest.cpp
@@ -159,6 +159,7 @@ TEST_F(ShuntingYardTest, willHandleSymbolAddition)
 
 TEST_F(ShuntingYardTest, willHandleValueAsignment)
 {
+    // x1 = 11.1
     InSequence s;
     nextId("x1");
     nextSym('=');
@@ -172,7 +173,6 @@ TEST_F(ShuntingYardTest, willHandleValueAsignment)
 
 TEST_F(ShuntingYardTest, willHandleValueAsignmentIncludingPrecedence)
 {
-
     // x1 = 14.7 + 4.1 * (1.5 - 0.7)
     InSequence s;
     nextId("x1");
@@ -193,32 +193,3 @@ TEST_F(ShuntingYardTest, willHandleValueAsignmentIncludingPrecedence)
     validateLispyForm(symbolTree.get(), "(= x1 (+ 14.7 (* 4.1 (- 1.5 0.7))))");
 }
 }
-
-#include <locale>
-#include <string>
-using namespace std::literals;
-
-// Declaration of the concept "EqualityComparable", which is satisfied by
-// any type T such that for values a and b of type T,
-// the expression a==b compiles and its result is convertible to bool
-// template <typename T>
-// concept bool EqualityComparable = requires(T a, T b)
-// {
-//     {
-//         a == b
-//     }
-//     ->bool;
-// };
-
-// template <EqualityComparable T>
-// bool f(T&& a)
-// {
-//     return a == T{};
-// }
-
-// TEST(dupa, dupa)
-// {
-//     std::cout << f("abc"s); // OK, std::string is EqualityComparable
-//     std::cout << f(std::use_facet<std::ctype<char>>(
-//         std::locale{})); // Error: not EqualityComparable
-// }
